requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6  # For file uploads

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pytest-xdist==3.5.0  # Parallel test execution

# Test data factories (optional)
factory-boy==3.3.0
faker==22.0.0

# Development tools
black==23.12.1
flake8==6.1.0
mypy==1.7.1
